pyproject.toml
[tool.poetry]
name = "dreaming-health-100"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.11.*"
streamlit = "^1.40.2"
pandas = "^2.2.3"
langchain = "^0.3.9"
langchain-cohere = "^0.3.3"
langchain-community = "^0.3.9"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
